Upgraded wngram2idngram to use long long int
wngram2idngram is now incompatible with original CMU-SLM's output
because it uses long long int to store counts instead of int.

New wngram2idngram will ignore individiual counts that don't fit int
limits (approx 2G)

+++ include/SLM2.h
@@ -402,7 +402,7 @@ results obtained from use of this software.
 
 typedef unsigned short id__t; /* Double underscore, since id_t is
 				 already defined on some platforms */
-typedef int count_t;   /* The count as read in, rather than its index 
+typedef long long int count_t;   /* The count as read in, rather than its index
 			  in the count table. */
 typedef unsigned short count_ind_t; /* The count's index in the count 
 				       table. */
